Move "browse by tipsiness" feature to subdirectory
