+++ frontend/src/searchform/searchform.tsx
@@ -3,7 +3,7 @@ import './searchform.css';
 
 import { IngredientList, loadAllIngredients } from './IngredientList';
 import { IIngredient } from "../../../types";
-import { useTipsySelector } from '../tipsyselector/tipsyselector';
+import { useTipsySelector } from '../tipsiness/tipsyselector/tipsyselector';
 
 const INGREDIENT_STORE = "get-me-drunk-fridge";
