+++ frontend/src/tipsiness/tipsyselector/math2d.ts
