+++ frontend/src/tipsiness/tipsyselector/tipsyselector.css
